backend/pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "MatiasPineda", email = "matias.pinedav@gmail.com"},
]
dependencies = [
    "Django>=4.2.5",
    "djangorestframework>=3.14.0",
    "djangorestframework-simplejwt>=5.3.0",
    "django-cors-headers>=4.2.0",
    "setuptools>=68.2.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

